pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lx06-flash-tool"
version = "0.1.0"
description = "All-in-One tool to manage, customize, and flash the Xiaomi LX06 (Xiaoai Speaker Pro)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "LX06 Flash Tool Contributors"},
]
keywords = ["xiaomi", "lx06", "firmware", "flash", "amlogic", "smart-speaker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console :: Curses",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    "textual>=0.40.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "docker>=6.0",
    "jinja2>=3.1",
    "httpx>=0.24",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "textual-dev>=0.20",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
lx06-tool = "lx06_tool.app:main"

[tool.setuptools.packages.find]
include = ["lx06_tool*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
